requirements.txt
flask==3.0.3
gunicorn==22.0.0
scikit-learn==1.5.1
datasets==2.20.0
google-auth==2.32.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.140.0
python-dotenv>=1.0.0
pandas>=2.0.0
torch>=2.0.0
transformers>=4.40.0,<5.0.0
sentencepiece>=0.1.99
ctranslate2>=3.24.0,<4.0.0
